finish the project

+++ 19T1021036/19T1021036.BusinessLayers/ProductDataService.cs
@@ -157,7 +157,7 @@ namespace _19T1021036.BusinessLayers
         /// </summary>
         /// <param name="attributeID"></param>
         /// <returns></returns>
-        public static ProductAttribute GetAttribute(int attributeID)
+        public static ProductAttribute GetAttribute(long attributeID)
         {
             return productDB.GetAttribute(attributeID);
         }
